move assist classes to new file -> closes q2g/q2g-ext-alternatestates#12
