+++ src/q2g-ext-altStateExtension.ts
@@ -1,4 +1,4 @@
-//#region
+//#region imports
 import * as qvangular from "qvangular";
 import * as qlik from "qlik";
 import * as template from "text!./q2g-ext-altStateExtension.html";
